refactored avatar module 2

+++ src/avatars/avatar.service.ts
@@ -9,53 +9,56 @@ export class AvatarsService {
     @Inject('StorageService') private storage: StorageService,
   ) {}
 
+  private async getAvatarUrl(userId: string) {
+    const user = await this.prisma.user.findUnique({
+      where: { id: userId },
+      select: { avatarUrl: true },
+    });
+    return user?.avatarUrl;
+  }
+
+  private async deleteAvatarIfExists(avatarUrl: string) {
+    if (avatarUrl) {
+      try {
+        await this.storage.delete(avatarUrl);
+      } catch (error) {
+        console.error('Error deleting avatar:', error);
+      }
+    }
+  }
+
   getImageURL(filename: string) {
     return this.storage.get(filename);
   }
 
   async upload(image: Express.Multer.File, userId: string) {
     const filename = image.originalname + Date.now();
-
-    const { avatarUrl } = await this.prisma.user.findUnique({
-      where: {
-        id: userId,
-      },
-    });
+    const avatarUrl = await this.getAvatarUrl(userId);
+    await this.deleteAvatarIfExists(avatarUrl);
     if (avatarUrl) {
       await this.storage.delete(avatarUrl);
     }
 
     try {
       await this.storage.upload(filename, image);
+      await this.prisma.user.update({
+        where: { id: userId },
+        data: { avatarUrl: filename },
+      });
     } catch (error) {
-      console.log(error);
+      throw new Error('Failed to upload avatar: ' + error.message);
     }
-    await this.prisma.user.update({
-      where: {
-        id: userId,
-      },
-      data: {
-        avatarUrl: filename,
-      },
-    });
 
     return this.getImageURL(filename);
   }
 
   async deleteByUserId(userId: string) {
-    const { avatarUrl } = await this.prisma.user.findUnique({
-      where: {
-        id: userId,
-      },
-    });
-    await this.storage.delete(avatarUrl);
+    const avatarUrl = await this.getAvatarUrl(userId);
+    await this.deleteAvatarIfExists(avatarUrl);
+
     await this.prisma.user.update({
-      where: {
-        id: userId,
-      },
-      data: {
-        avatarUrl: null,
-      },
+      where: { id: userId },
+      data: { avatarUrl: null },
     });
   }
 }
